add memory size check

+++ gen_tea_id.sh
@@ -2,6 +2,12 @@
 
 TEA_CONFIG="$HOME/.tea"
 
+MEM_SIZE=`grep MemTotal /proc/meminfo | awk '{printf "%.0f", ($2 / 1024)}'`
+if [ "$MEM_SIZE" -lt 1800 ]; then
+  error "Machine memory size should larger equal than 2G"
+  exit 1
+fi
+
 sudo apt-get install -y wget
 
 rm -f mock-tea-id-generator
